python/requirements.txt
future>=1.0.0
numpy>=2.1.1
opencv-python>=4.10.0.84
pillow>=10.4.0
